move extrasolar stuff into "generic behaviour"
